requirements.txt
Flask~=3.0.3
pymongo~=4.6.3
numpy~=1.26.4
tensorflow~=2.16.1
requests~=2.31.0
pika~=1.3.2
python-dotenv~=1.0.1
keras~=3.2.1
tf-nightly
scikit-learn~=1.4.2
pillow~=10.3.0